refactor: merge commands/remove package into commands/setup
A single-file top-level folder for one teardown command added a
directory for no organizational benefit. CLI group/name (bench remove
production) is unchanged, only the source file moved.
