+++ pilot/commands/setup/remove_production.py
